deprecated/requirements.txt
# GA4 API 관련
google-analytics-admin>=0.21.0
google-auth>=2.0.0
google-auth-oauthlib>=0.7.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# 웹 프레임워크
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# 데이터베이스
aiosqlite>=0.19.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# 스케줄링 및 비동기
schedule==1.2.0
apscheduler>=3.10.0

# 유틸리티
python-dotenv>=1.0.0
pydantic>=2.0.0
python-dateutil>=2.8.0
email-validator>=2.0.0

# 로깅 및 모니터링
structlog>=23.0.0
rich>=13.0.0

# 개발 도구
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0 